iOS/Mac compilation

+++ Framework/3rdPartyLibs/Box2D/Common/b2Math.h
@@ -46,8 +46,8 @@ inline float32 b2InvSqrt(float32 x)
 	return x;
 }
 
-#define	b2Sqrt(x)	sqrtf(x)
-#define	b2Atan2(y, x)	atan2f(y, x)
+#define	b2Sqrt(x)	sqrt(x)
+#define	b2Atan2(y, x)	atan2(y, x)
 
 /// A 2D column vector.
 struct b2Vec2
@@ -304,16 +304,16 @@ struct b2Rot
 	explicit b2Rot(float32 angle)
 	{
 		/// TODO_ERIN optimize
-		s = sinf(angle);
-		c = cosf(angle);
+		s = sin(angle);
+		c = cos(angle);
 	}
 
 	/// Set using an angle in radians.
 	void Set(float32 angle)
 	{
 		/// TODO_ERIN optimize
-		s = sinf(angle);
-		c = cosf(angle);
+		s = sin(angle);
+		c = cos(angle);
 	}
 
 	/// Set to the identity rotation
@@ -712,7 +712,7 @@ inline void b2Sweep::Advance(float32 alpha)
 inline void b2Sweep::Normalize()
 {
 	float32 twoPi = 2.0f * b2_pi;
-	float32 d =  twoPi * floorf(a0 / twoPi);
+	float32 d =  twoPi * floor(a0 / twoPi);
 	a0 -= d;
 	a -= d;
 }
